+++ plugins/services/src/js/reducers/JSONParserReducers.js
@@ -1,7 +1,7 @@
 import {simpleParser} from '../../../../../src/js/utils/ParserUtil';
 
-import {JSONParser as environmentVariables} from './form/EnvironmentVariables';
-import {JSONParser as labels} from './form/Labels';
+import {JSONParser as environmentVariables} from './serviceForm/EnvironmentVariables';
+import {JSONParser as labels} from './serviceForm/Labels';
 
 module.exports = [
   simpleParser(['id']),
